Break classes in two files, using attr reader for main.rb and change require to require_relative for internal libraries
